New Runge Kutta 4 file, unfinished.
Unfinished Runge Kutta 4 solution Jupyter Notebook, solving a probe particle falling into a black hole, using General Relativity Equations. 

Notes on how to solve the problem are in a part of the file, the described algorithm just has to be updated for the calculation of the GR formulas used to describe the motion of the probe particle close to a black hole. Different cases can then be simulated and displayed easily, by just changing the start params of the eom's. 

Idea to fit alg to equations: Update rules known for three-parameter runge kutta for dashed differentials, calculate undashed differentials regularly. Implement dashed differentials formulations as functions, that return the equation.